+++ Tools/RMINTGenericPropRigTool.py
@@ -25,7 +25,7 @@ from RMPY.snippets import NoisePatternRig
 
 def getMayaWindow():
     ptr = mui.MQtUtil.mainWindow()
-    return wrapInstance(long(ptr), QMainWindow)
+    return wrapInstance(int(ptr), QMainWindow)
 
 class main(MayaQWidgetDockableMixin, QDialog):
     def __init__(self, NameConv=None, parent=None):
